fix: add a way to update the instances and avoiding relationals errors
