+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "db:seed": "node ./database/mocks/seed.js"
+    "db:seed": "node ./database/mocks/create/seed.js && node ./database/mocks/update/update.js",
   },
   "repository": {
     "type": "git",
